+++ src/ZeppelinForms/Forms/Controls/CheckBox.cs
@@ -59,8 +59,8 @@ public class CheckBox : InteractiveControl, ITextElement
     public Color BoxBorderColor { get; set; } = Colors.Black;
     public Color BoxBackground { get; set; } = Colors.White;
     public Color CheckColor { get; set; } = new Color(255, 0x0D, 0x6E, 0xFD);
-    public HorizontalContentAlignment HorizontalContentAlign { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-    public VerticalContentAlignment VerticalContentAlign { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+    public HorizontalContentAlignment HorizontalContentAlign { get; set; }
+    public VerticalContentAlignment VerticalContentAlign { get; set; }
 
     public CheckBox()
     {